+++ assets/javascripts/kitten/components/steppers/stepper-icon.js
@@ -14,7 +14,9 @@ const StepperIcon = ({ className, children }) => {
   )
 
   return (
-    <span className={ iconClassNames }>{ children }</span>
+    <span className={ iconClassNames }>
+      { children }
+    </span>
   )
 }
 
@@ -25,9 +27,15 @@ export const StepperIconDefault = () => {
 }
 
 export const StepperIconInProgress = () => {
+  const inProgressClassNames = classNames(
+    'k-Stepper__icon--inProgress',
+  )
+
   return (
-    <StepperIcon className="k-Stepper__icon--inProgress">
-      <ArrowIcon className="k-ButtonIcon__svg k-Stepper__svg" />
+    <StepperIcon className={ inProgressClassNames }>
+      <ArrowIcon className={ classNames('k-ButtonIcon__svg',
+                                        'k-Stepper__svg') }
+      />
     </StepperIcon>
   )
 }
@@ -40,7 +48,9 @@ export const StepperIconValidated = () => {
 
   return (
     <StepperIcon className={ validatedClassNames }>
-      <CheckedIcon className="k-ButtonIcon__svg k-Stepper__svg" />
+      <CheckedIcon className={ classNames('k-ButtonIcon__svg',
+                                          'k-Stepper__svg') }
+      />
     </StepperIcon>
   )
 }